Restructuring files, (error in tic-tac-toe user.js)
