+++ package.json
@@ -45,4 +45,3 @@
     "nodemon": "^1.18.11"
   }
 }
-
