+++ src/games.js
@@ -1,1 +1,2 @@
 require("./games/tic-tac-toe")
+require("./games/chain-reaction")
